Replcae useless extra float check in StringUtils::cast with double
C++ doesn't have an unsigned float type and if it did, this isn't using it

It does have doubles though, and both _wtof and std::atof return those.

+++ include/Utils/StringUtils.hpp
@@ -270,7 +270,7 @@ class StringUtils
                     return std::atoll(str.data());
                 }
             }
-            else if constexpr (std::is_same_v<Ret, float> || std::is_same_v<Ret, float>)
+            else if constexpr (std::is_same_v<Ret, float> || std::is_same_v<Ret, double>)
             {
                 if constexpr (IsWide)
                 {
